backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
gunicorn>=22.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.0.1
python-multipart>=0.0.9
email-validator>=2.1
httpx>=0.27
reportlab>=4.0
num2words>=0.5
pandas>=2.2
openpyxl>=3.1
pyotp>=2.9
qrcode>=7.4
pytest>=8.0
# Observability
opentelemetry-api>=1.23
opentelemetry-sdk>=1.23
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-sqlalchemy>=0.44b0
opentelemetry-exporter-otlp>=1.23
prometheus-client>=0.20
